fix(parallel-map): flush implementation
In case _flush method is called when there are running transformation,
the user flush call has to be postponed waiting for the transformations
to be completed.
On the other hand, in case the _flush method is called when no transformation
is pending, the user flush method has to be called immediately.